Report Uncompressed Size of archive for embed...
ed zip files
At least in TDMF's case, the game needs to know the proper uncompressed size of every zip archive, we now account for this by serializing zip archive data and counting up total size of uncompressed data insize a zip file.
std.zip cannot be used because it reports the zip files as malformed, so a separate library was used.